+++ apps/web/app/(pages)/(public)/(logged-out-only)/page.tsx
@@ -1,11 +1,16 @@
 import React from "react";
 
-import GeneralLanding from "@good-dog/components/landing-pages/GeneralLanding";
+import GeneralLanding from "@good-dog/components/landing-pages/general/GeneralLanding";
 import PageContainer from "@good-dog/components/PageContainer";
 
 export default async function Home() {
   return (
-    <PageContainer background="gradient" widthType="small" allowMobile={true}>
+    <PageContainer
+      background="gradient"
+      widthType="full"
+      allowMobile={true}
+      smallHeader={true}
+    >
       <GeneralLanding />
     </PageContainer>
   );
